refactor: move the phase kinds back into phases, only keep adaptive scale seperate
