+++ src/aiperf/config/adaptive_scale_phase.py
@@ -4,6 +4,11 @@
 
 from __future__ import annotations
 
+from typing import Annotated, Literal
+
+from pydantic import Field, model_validator
+from typing_extensions import Self
+
 from aiperf.config.sweep.adaptive import SLAFilter
 
 
@@ -99,3 +104,148 @@ def lower_adaptive_scale_details(
         lowered["sla"] = sla
     elif isinstance(sla, dict):
         lowered["sla"] = normalize_adaptive_sla(sla)
+
+
+class AdaptiveScalePhaseMixin:
+    """Adaptive scale fields and validation for concurrency phases."""
+
+    adaptive_scale: Annotated[
+        bool,
+        Field(
+            default=False,
+            description="Enable single-run adaptive scale control for this phase.",
+        ),
+    ]
+
+    adaptive_sustain_duration: Annotated[
+        float | None,
+        Field(
+            gt=0,
+            default=None,
+            description="Duration in seconds to sustain load near the discovered adaptive scale boundary.",
+        ),
+    ]
+
+    adaptive_assessment_period: Annotated[
+        float | None,
+        Field(
+            ge=1.0,
+            default=None,
+            description="Duration in seconds for each adaptive scale SLA assessment window.",
+        ),
+    ]
+
+    adaptive_min_completed_requests: Annotated[
+        int,
+        Field(
+            ge=1,
+            default=1,
+            description="Minimum completed requests needed before an adaptive SLA window can make a decision.",
+        ),
+    ]
+
+    adaptive_control_variable: Annotated[
+        Literal["concurrency"],
+        Field(
+            default="concurrency",
+            description="Named adaptive control variable. Only concurrency is supported in v1.",
+        ),
+    ]
+
+    adaptive_scale_min_concurrency: Annotated[
+        int,
+        Field(
+            ge=1,
+            default=1,
+            description="Minimum concurrency used by adaptive scale discovery.",
+        ),
+    ]
+
+    adaptive_scale_strategy_type: Annotated[
+        Literal["ramp_until_fail"],
+        Field(
+            default="ramp_until_fail",
+            description="Adaptive scale controller strategy. v1 supports ramp_until_fail.",
+        ),
+    ]
+
+    adaptive_scale_step_policy: Annotated[
+        Literal["sla_margin", "fixed_percent_step"],
+        Field(
+            default="sla_margin",
+            description=(
+                "Adaptive scale increase policy. sla_margin uses normalized SLA "
+                "margin to choose larger steps when far from the boundary; "
+                "fixed_percent_step uses a fixed percentage of the current control value."
+            ),
+        ),
+    ]
+
+    adaptive_scale_base_step: Annotated[
+        int,
+        Field(
+            ge=1,
+            default=10,
+            description="Minimum adaptive scale step for SLA-margin policy.",
+        ),
+    ]
+
+    adaptive_scale_max_step_multiplier: Annotated[
+        int,
+        Field(
+            ge=1,
+            default=4,
+            description="Maximum base-step multiplier for SLA-margin policy.",
+        ),
+    ]
+
+    adaptive_scale_step_percent: Annotated[
+        float,
+        Field(
+            gt=0,
+            default=25.0,
+            description="Percent of current concurrency used by fixed-percent adaptive scaling.",
+        ),
+    ]
+
+    @model_validator(mode="before")
+    @classmethod
+    def _lower_adaptive_scale_block(cls, data: object) -> object:
+        if not isinstance(data, dict):
+            return data
+        lowered = dict(data)
+        if isinstance(lowered.get("sla"), dict):
+            lowered["sla"] = normalize_adaptive_sla(lowered["sla"])
+
+        block = data.get("adaptive_scale")
+        if not isinstance(block, dict):
+            return lowered
+
+        lower_adaptive_scale_details(lowered, block)
+        return lowered
+
+    @model_validator(mode="after")
+    def _validate_adaptive_scale(self) -> Self:
+        if not self.adaptive_scale:
+            return self
+        if self.duration is None:
+            raise ValueError("adaptive_scale requires duration")
+        if self.adaptive_sustain_duration is None:
+            raise ValueError("adaptive_scale requires adaptive_sustain_duration")
+        if not self.sla:
+            raise ValueError("adaptive_scale requires sla filters")
+        if self.concurrency_ramp is not None:
+            raise ValueError(
+                "adaptive_scale cannot be combined with concurrency_ramp. "
+                "adaptive_scale already adjusts concurrency during the phase to "
+                "discover an SLA boundary. Use concurrency_ramp only when you know "
+                "the target concurrency and want to ease into it over a fixed duration."
+            )
+        # TODO: AIP-967 - Add adaptive scale control-backend abstraction.
+        if self.adaptive_control_variable != "concurrency":
+            raise ValueError(
+                "adaptive_scale control variable must be 'concurrency' in this release"
+            )
+        if self.adaptive_scale_min_concurrency > self.concurrency:
+            raise ValueError("adaptive_scale_min_concurrency must be <= concurrency")
+        return self
