+++ src/aiperf/config/phases.py
@@ -1,7 +1,5 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
-# ruff: noqa: I001
-
 """
 AIPerf Configuration v2.0 - Phase Configuration
 
@@ -22,12 +20,9 @@ from pydantic import (
 )
 from typing_extensions import Self
 
+from aiperf.config.adaptive_scale_phase import AdaptiveScalePhaseMixin
 from aiperf.config.base import BaseConfig
 from aiperf.config.cancellation import CancellationConfig
-from aiperf.config.adaptive_scale_phase import (
-    lower_adaptive_scale_details,
-    normalize_adaptive_sla,
-)
 from aiperf.config.loader.duration import (
     DurationSpec,
     _normalize_duration,
@@ -282,7 +277,7 @@ class BasePhaseConfig(BaseConfig):
 # =============================================================================
 
 
-class ConcurrencyPhase(BasePhaseConfig):
+class ConcurrencyPhase(AdaptiveScalePhaseMixin, BasePhaseConfig):
     """Concurrency-controlled load: dispatch immediately when a slot opens.
 
     Primary control is ``concurrency`` (defaults to 1).
@@ -304,160 +299,161 @@ class ConcurrencyPhase(BasePhaseConfig):
         ),
     ]
 
-    adaptive_scale: Annotated[
-        bool,
-        Field(
-            default=False,
-            description="Enable single-run adaptive scale control for this phase.",
-        ),
-    ]
-
-    adaptive_sustain_duration: Annotated[
-        float | None,
-        Field(
-            gt=0,
-            default=None,
-            description="Duration in seconds to sustain load near the discovered adaptive scale boundary.",
-        ),
-    ]
-
-    adaptive_assessment_period: Annotated[
-        float | None,
-        Field(
-            ge=1.0,
-            default=None,
-            description="Duration in seconds for each adaptive scale SLA assessment window.",
-        ),
-    ]
-
-    adaptive_min_completed_requests: Annotated[
-        int,
-        Field(
-            ge=1,
-            default=1,
-            description="Minimum completed requests needed before an adaptive SLA window can make a decision.",
-        ),
-    ]
-
-    adaptive_control_variable: Annotated[
-        Literal["concurrency"],
-        Field(
-            default="concurrency",
-            description="Named adaptive control variable. Only concurrency is supported in v1.",
-        ),
-    ]
-
-    adaptive_scale_min_concurrency: Annotated[
-        int,
-        Field(
-            ge=1,
-            default=1,
-            description="Minimum concurrency used by adaptive scale discovery.",
-        ),
-    ]
-
-    adaptive_scale_strategy_type: Annotated[
-        Literal["ramp_until_fail"],
-        Field(
-            default="ramp_until_fail",
-            description="Adaptive scale controller strategy. v1 supports ramp_until_fail.",
-        ),
-    ]
-
-    adaptive_scale_step_policy: Annotated[
-        Literal["sla_margin", "fixed_percent_step"],
-        Field(
-            default="sla_margin",
-            description=(
-                "Adaptive scale increase policy. sla_margin uses normalized SLA "
-                "margin to choose larger steps when far from the boundary; "
-                "fixed_percent_step uses a fixed percentage of the current control value."
-            ),
-        ),
-    ]
-
-    adaptive_scale_base_step: Annotated[
-        int,
-        Field(
-            ge=1,
-            default=10,
-            description="Minimum adaptive scale step for SLA-margin policy.",
-        ),
-    ]
-
-    adaptive_scale_max_step_multiplier: Annotated[
-        int,
-        Field(
-            ge=1,
-            default=4,
-            description="Maximum base-step multiplier for SLA-margin policy.",
-        ),
-    ]
-
-    adaptive_scale_step_percent: Annotated[
-        float,
-        Field(
-            gt=0,
-            default=25.0,
-            description="Percent of current concurrency used by fixed-percent adaptive scaling.",
-        ),
-    ]
-
-    @model_validator(mode="before")
-    @classmethod
-    def _lower_adaptive_scale_block(cls, data: object) -> object:
-        if not isinstance(data, dict):
-            return data
-        lowered = dict(data)
-        if isinstance(lowered.get("sla"), dict):
-            lowered["sla"] = normalize_adaptive_sla(lowered["sla"])
-
-        block = data.get("adaptive_scale")
-        if not isinstance(block, dict):
-            return lowered
-
-        lower_adaptive_scale_details(lowered, block)
-        return lowered
-
-    @model_validator(mode="after")
-    def _validate_adaptive_scale(self) -> Self:
-        if not self.adaptive_scale:
-            return self
-        if self.duration is None:
-            raise ValueError("adaptive_scale requires duration")
-        if self.adaptive_sustain_duration is None:
-            raise ValueError("adaptive_scale requires adaptive_sustain_duration")
-        if not self.sla:
-            raise ValueError("adaptive_scale requires sla filters")
-        if self.concurrency_ramp is not None:
-            raise ValueError(
-                "adaptive_scale cannot be combined with concurrency_ramp. "
-                "adaptive_scale already adjusts concurrency during the phase to "
-                "discover an SLA boundary. Use concurrency_ramp only when you know "
-                "the target concurrency and want to ease into it over a fixed duration."
-            )
-        # TODO: AIP-967 - Add adaptive scale control-backend abstraction.
-        if self.adaptive_control_variable != "concurrency":
-            raise ValueError(
-                "adaptive_scale control variable must be 'concurrency' in this release"
-            )
-        if self.adaptive_scale_min_concurrency > self.concurrency:
-            raise ValueError("adaptive_scale_min_concurrency must be <= concurrency")
-        return self
-
 
 # =============================================================================
 # RATE-CONTROLLED PHASES
 # =============================================================================
 
-from aiperf.config.phase_kinds import (  # noqa: E402
-    ConstantPhase,
-    FixedSchedulePhase,
-    GammaPhase,
-    PoissonPhase,
-    RatePhaseConfig,
-    UserCentricPhase,
-)
+
+class RatePhaseConfig(BasePhaseConfig):
+    """Base for rate-controlled phases. Not instantiated directly."""
+
+    rate: Annotated[
+        float,
+        Field(
+            gt=0,
+            description="Target request rate in requests per second (must be > 0).",
+        ),
+    ]
+
+    rate_ramp: Annotated[
+        RampSpec,
+        Field(
+            default=None,
+            description="Ramp rate from lower value. "
+            "Can be number (seconds) or {duration, strategy}.",
+        ),
+    ]
+
+
+class PoissonPhase(RatePhaseConfig):
+    """Poisson-distributed request arrivals at the target rate."""
+
+    type: Annotated[
+        Literal[PhaseType.POISSON],
+        Field(description="Poisson-distributed rate-controlled arrivals."),
+    ]
+
+
+class GammaPhase(RatePhaseConfig):
+    """Gamma-distributed request arrivals with configurable smoothness."""
+
+    type: Annotated[
+        Literal[PhaseType.GAMMA],
+        Field(description="Gamma-distributed rate-controlled arrivals."),
+    ]
+
+    smoothness: Annotated[
+        float | None,
+        Field(
+            gt=0,
+            default=None,
+            description="Gamma distribution shape parameter (must be > 0). "
+            "1.0 = Poisson, <1 = bursty, >1 = regular.",
+        ),
+    ]
+
+
+class ConstantPhase(RatePhaseConfig):
+    """Constant-rate request arrivals (fixed inter-arrival time)."""
+
+    type: Annotated[
+        Literal[PhaseType.CONSTANT],
+        Field(description="Constant rate-controlled arrivals."),
+    ]
+
+
+class UserCentricPhase(RatePhaseConfig):
+    """N concurrent users sharing a global request rate.
+
+    Requires multi-turn conversations. Each user gets a proportional
+    share of the global ``rate``.
+    """
+
+    type: Annotated[
+        Literal[PhaseType.USER_CENTRIC],
+        Field(description="N users sharing a global request rate."),
+    ]
+
+    users: Annotated[
+        int,
+        Field(
+            ge=1,
+            description="Number of simulated concurrent users (must be >= 1). "
+            "Requests distributed across users to achieve global rate.",
+        ),
+    ]
+
+    @model_validator(mode="after")
+    def validate_user_centric_constraints(self) -> UserCentricPhase:
+        """Validate user-centric mode constraints."""
+        if self.sessions is not None and self.sessions < self.users:
+            raise ValueError(
+                f"Phase '{self.name}': --num-sessions ({self.sessions}) must be "
+                f">= --num-users ({self.users}). Each user needs at least one session."
+            )
+
+        if self.requests is not None and self.requests < self.users:
+            raise ValueError(
+                f"Phase '{self.name}': --request-count ({self.requests}) must be "
+                f">= --num-users ({self.users}). Each user needs at least one request."
+            )
+
+        return self
+
+
+class FixedSchedulePhase(BasePhaseConfig):
+    """Replay requests at predetermined timestamps from a trace dataset.
+
+    Stop condition not required -- the trace dataset determines when the
+    phase ends.
+    """
+
+    _stop_condition_required: ClassVar[bool] = False
+
+    type: Annotated[
+        Literal[PhaseType.FIXED_SCHEDULE],
+        Field(description="Replay requests at trace timestamps."),
+    ]
+
+    auto_offset: Annotated[
+        bool,
+        Field(
+            default=True,
+            description="Normalize trace timestamps to start at 0. "
+            "Subtracts minimum timestamp from all entries.",
+        ),
+    ]
+
+    start_offset: Annotated[
+        int | None,
+        Field(
+            ge=0,
+            default=None,
+            description="Filter out trace requests before this timestamp in ms (must be >= 0).",
+        ),
+    ]
+
+    end_offset: Annotated[
+        int | None,
+        Field(
+            ge=0,
+            default=None,
+            description="Filter out trace requests after this timestamp in ms (must be >= 0).",
+        ),
+    ]
+
+    @model_validator(mode="after")
+    def _validate_fixed_schedule_constraints(self) -> Self:
+        if self.auto_offset and self.start_offset is not None:
+            raise ValueError("auto_offset cannot be True when start_offset is set")
+        if (
+            self.start_offset is not None
+            and self.end_offset is not None
+            and self.start_offset > self.end_offset
+        ):
+            raise ValueError("start_offset must be <= end_offset")
+        return self
 
 
 # =============================================================================
